backend/requirements.txt
# backend/requirements.txt
google-adk
google-generativeai
python-dotenv
google-api-python-client
google-auth-oauthlib
fastapi
uvicorn